Separate out swap input output currencies
